requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
passlib==1.7.4
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.4
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
starlette-session==0.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
